Move factories to dedicated package
