+++ jparest/src/main/java/uk/gov/homeoffice/digital/sas/jparest/config/ApplicationConfig.java
@@ -9,11 +9,11 @@ import org.springframework.context.annotation.Import;
 import org.springframework.context.support.GenericApplicationContext;
 import org.springframework.transaction.PlatformTransactionManager;
 import uk.gov.homeoffice.digital.sas.jparest.ResourceEndpoint;
-import uk.gov.homeoffice.digital.sas.jparest.controller.ResourceApiControllerFactory;
 import uk.gov.homeoffice.digital.sas.jparest.exceptions.exceptionhandling.ApiResponseExceptionHandler;
+import uk.gov.homeoffice.digital.sas.jparest.factory.ResourceApiControllerFactory;
+import uk.gov.homeoffice.digital.sas.jparest.factory.ResourceApiServiceFactory;
 import uk.gov.homeoffice.digital.sas.jparest.service.BaseEntityCheckerService;
 import uk.gov.homeoffice.digital.sas.jparest.service.ControllerRegistererService;
-import uk.gov.homeoffice.digital.sas.jparest.service.ResourceApiServiceFactory;
 import uk.gov.homeoffice.digital.sas.jparest.swagger.PathItemCreator;
 import uk.gov.homeoffice.digital.sas.jparest.swagger.ResourceOpenApiCustomiser;
 import uk.gov.homeoffice.digital.sas.jparest.validation.EntityValidator;
