+++ jparest/src/test/java/uk/gov/homeoffice/digital/sas/jparest/factory/ResourceApiControllerFactoryTest.java
@@ -1,4 +1,4 @@
-package uk.gov.homeoffice.digital.sas.jparest.controller;
+package uk.gov.homeoffice.digital.sas.jparest.factory;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.support.GenericApplicationContext;
 import org.springframework.test.context.ContextConfiguration;
+import uk.gov.homeoffice.digital.sas.jparest.controller.ResourceApiController;
 import uk.gov.homeoffice.digital.sas.jparest.entityutils.testentities.DummyEntityA;
 import uk.gov.homeoffice.digital.sas.jparest.service.ResourceApiService;
